nanobot-video-generator/videogenerator/pyproject.toml
[project]
name = "mcp-server-videogenerator"
version = "0.0.3"
description = "A Model Context Protocol server providing tools to generate videos"
requires-python = ">=3.13"
authors = [{ name = "drpebcak" }]
keywords = ["http", "mcp", "llm", "automation"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "fastmcp>=2.6.1",
    "google-genai>=1.19.0",
    "google-auth",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
]

[project.scripts]
mcp-server-videogenerator = "mcp_server_videogenerator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.11.12"]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_videogenerator"]
